pyproject.toml
[tool.poetry]
name = "template-ecommerce-django"
version = "0.1.0"
description = ""
authors = ["foschieraanderson <foschieraanderson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "2.8.6"
whitenoise = "5.2"
gunicorn = "^20.1.0"
django-localflavor = "^3.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
